backend/requirements.txt
openclaw-sdk>=0.5.0
ethoswarm-sdk>=1.2.0
openai>=1.0.0
web3>=6.0.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.24
python-dotenv>=1.0
sentence-transformers>=2.2.0
redis>=4.0.0
eth-account>=0.10.0
